# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pauli API
  description: |
    Read-only HTTP API for validator data indexed by Pauli into PostgreSQL.
    All list responses use a `data` array and `meta` object with `limit`, `offset`, and `count`
    (number of rows returned in this page).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local pauli-api
paths:
...
  /v1/validators/{validatorIndex}/snapshots/latest:
    get:
      summary: Latest snapshot for a validator
      operationId: getLatestSnapshot
      parameters:
        - $ref: "#/components/parameters/validatorIndexPath"
      responses:
        "200":
          description: Snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorSnapshot"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/validators/{validatorIndex}/snapshots:
    get:
      summary: Snapshots for a validator in a slot range
      operationId: listSnapshots
      parameters:
        - $ref: "#/components/parameters/validatorIndexPath"
        - name: from_slot
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: to_slot
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorSnapshotListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/validators/{validatorIndex}/snapshots/count:
    get:
      summary: Number of snapshot rows for a validator
      operationId: countSnapshots
      parameters:
        - $ref: "#/components/parameters/validatorIndexPath"
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/validators/{validatorIndex}/penalties:
    get:
      summary: Validator penalties in an epoch range
      operationId: listPenalties
      parameters:
        - $ref: "#/components/parameters/validatorIndexPath"
        - name: epoch
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: from_epoch
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: to_epoch
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated penalties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorPenaltyListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Page size (default 100, maximum 1000)
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    offset:
      name: offset
      in: query
      description: Number of rows to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    validatorIndexPath:
      name: validatorIndex
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9]+$"
...
  responses:
    BadRequest:
      description: Invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    ListMeta:
      type: object
      required: [limit, offset, count]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        count:
          type: integer
          description: Number of items in `data` for this response
    ValidatorSnapshot:
      type: object
      properties:
        validator_index:
          type: integer
          format: int64
        slot:
          type: integer
          format: int64
        status:
          type: string
        balance:
          type: integer
          format: int64
        effective_balance:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
    ValidatorSnapshotListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ValidatorSnapshot"
        meta:
          $ref: "#/components/schemas/ListMeta"
...
    ValidatorPenalty:
      type: object
      properties:
        validator_index:
          type: integer
          format: int64
        epoch:
          type: integer
          format: int64
        slot:
          type: integer
          format: int64
        penalty_type:
          type: string
        penalty_gwei:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
    ValidatorPenaltyListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ValidatorPenalty"
        meta:
          $ref: "#/components/schemas/ListMeta"